feat!: idiomatic generated types -- u64/Decimal/NaiveDate params, snake_case idents
The batch of remaining codegen-review items, all breaking changes to the
generated surface bundled into one regeneration:

* Integer request params are u64 (was i64), matching the response side.
  Audited: no integer param has negative semantics (the documented -1
  sentinels are enum-typed), and response ids are already u64, so i64
  forced a cast on every get/set round-trip -- the livetest harness shed
  ~30 `as i64` bridge casts.
* Decimal request params are rust_decimal::Decimal (was f64): charge,
  payment, setup, monthly, minute -- money amounts that f64 would
  serialize with float artifacts -- plus timezone. chrono gains its
  `serde` feature for the date params below.
* date_from/date_to params are chrono::NaiveDate via a new DATE_FIELDS
  builtin; NaiveDate's own Serialize emits the documented YYYY-MM-DD.
* Generated field idents are snake_case with a serde rename back to the
  wire name (isMobile -> is_mobile, sipuri -> sip_uri; 44 renames), on
  both the params and response side. #![allow(non_snake_case)] is gone.
  Params and responses share one keyword-escaping ident helper, closing
  a latent gap where a param named `match` would emit invalid Rust.
* addLNPPort's locationType (and getLNPDetails's matching response
  field) is the new LocationType enum (Residential/Business, wire 0/1).

Generator structure:

* A shared field_overrides::Resolver replaces the three-way override
  precedence duplicated between main.rs and response_codegen.rs, and
  adds a scalar-only guard: name-based substitutions never apply to
  collection-shaped response fields, so the getNAT/getPlayInstructions
  reference catalogs keep their structural types and their two
  field_type_skip entries are removed.
* New `cargo xtask check-flags` audits FLAG_01_FIELDS/FLAG_YES_NO_FIELDS
  against the doc-mined parameter descriptions (uncovered flag-like
  params and stale entries). Its first run caught locationType, which is
  typed as an enum rather than a flag. Documented as step 4 of the
  refresh procedure in DEVELOPMENT.md.

Three new wiremock tests pin the wire forms: charge=1.25 exact,
date_from=2026-01-01, and isMobile=1 (rename + flag serializer
composing on one field).

+++ livetest/src/areas/dids.rs
@@ -323,7 +323,7 @@ async fn order_fixture_did(ctx: &AreaCtx<'_>, report: &mut Report) {
         .order_did(&OrderDIDParams {
             did: Some(did.clone()),
             routing: Some(Routing::None),
-            pop: Some(pop as i64),
+            pop: Some(pop),
             dialtime: Some(60),
             cnam: Some(0),
             billing_type: Some(DidBillingType::PerMinute),
